Clean up End of Year code
