--- modules/features/endofyear/src/main/java/au/com/shiftyjelly/pocketcasts/endofyear/ui/Fonts.kt
@@ -1,10 +1,0 @@
-package au.com.shiftyjelly.pocketcasts.endofyear.ui
-
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import au.com.shiftyjelly.pocketcasts.endofyear.R
-
-internal val humaneFontFamily = FontFamily(
-    Font(R.font.humane_regular, FontWeight.Normal),
-)
